Added Segmented_string_compare and compatible_trains
Segmented_string_compare:
A function that takes two string where the input is a string from train and the other is target which comes from route.controls. It compares each word in the dot sperated to the target.

compatible_trains:
Compatible_trains is a function that takes the train struct pointer, a route and the train_count. It goes through each train in a for loop and uses Segmented_string_compare to detemine if the train can drive on the route